pyproject.toml
[project]
name = "taillog_labeling"
version = "0.1.0"
description = "YouTube dog behavior labeling pipeline"
requires-python = ">=3.11"

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "yt-dlp>=2023.12.0",
    "opencv-python>=4.8.0",
    "ultralytics>=8.0.0",
    "ollama>=0.1.0",
    "supabase>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
